ChatWithDoc/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
langchain
langchain-community
langchain-text-splitters
langchain-openai
openai
chromadb
pypdf
numpy
httpx
pytest
python-dotenv
lark
